initial Html Css Update
Changed the css and the Html file names. Added the links to the validate css and the forms css. Added the javascript link. Added the login form. Added the inputs for the username and the password. Added the submit button for the user and password name. Added the payment form. Added the legend of Billing information(required). Added the inputs and the labels for the data fields. Added the data list of all the initials of the states. Added an fieldset for the credit card information with the legend of CreditCard(required).