Html Update
Added the field set for the credit card information. Added another field set for type of credit cards. Added the labels and the inputs for the credit cards. Added the images of the different types of credit cards that are excepted.  Added the expiration date field. Added the csc field with an 3 character limit. Added the next submit button.