Final commit HTML Css update
For every input that is not a radio and that is focused will have the background color changed to a tan.  For every input that is not a radio and that is focused and that is valid  will have the background color changed to a green and will have a green valid logo at the right side show up. For every input that is not a radio and that is focused and that is invalid  will have the background color changed to a red and will have a red invalid logo at the right side show up. Added Comments for both html and css files.  Ready for review.